Fix bug in Encoder implementations
The Encoder trait requires that new messages be appended to the end of the
provided BytesMut. The cookie-factory serializers require that the underlying
buffer have pre-allocated space. In combining these two, it was mistakenly
assumed that the provided BytesMut may have non-zero size, but would be
completely empty (so serialization should happen from position 0).

This commit fixes the Encoder implementations to not overwrite any data already
in the BytesMut.